Replace internal API routes with server actions
